Backend/requirements.txt
fastapi
uvicorn
requests
scikit-learn
pytest
python-dotenv
passlib[bcrypt]
python-jose
pydantic[email]
python-multipart